+++ polybar/launch.sh
@@ -1,6 +1,10 @@
-#!/usr/bin/env bash
+#!/usr/bin/bash
 
 DIR="$HOME/.config/polybar"
-killall -q polybar
-while pgrep -u $UID -x polybar >/dev/null; do sleep 1; done
+
+if [[ $(pgrep -x polybar) ]]; then
+	killall -q polybar
+	while pgrep -x polybar >/dev/null; do sleep 1; done
+fi
+
 polybar -q main -c "$DIR"/config.ini &
